refactor: seperating teams from game schedule
