+++ src/components/matchList/gameSchedule/index.tsx
@@ -3,60 +3,7 @@ import * as S from "./styled";
 import { colors } from "../../../../styles";
 import { StatusContainer } from "../../common/StatusContainer";
 import { MatchStatus } from "../../common/LiveStatus";
-
-interface TeamProps {
-  teamData: {
-    department: string;
-    ratio: number;
-    class: string;
-    score: number;
-  };
-  color: string;
-  alignItems: string;
-  isDuring: boolean;
-  isEnd: boolean;
-}
-
-const Team: React.FC<TeamProps> = ({
-  teamData,
-  color,
-  alignItems,
-  isDuring,
-  isEnd,
-}) => {
-  return (
-    <S.TeamBox color={color} alignItems={alignItems}>
-      {isDuring ? (
-        <>
-          <S.ClassBox>
-            <S.Department>{teamData.department}</S.Department>
-            <S.Class fontSize={13}>{teamData.class}</S.Class>
-          </S.ClassBox>
-
-          <S.Ratio>{teamData.ratio}%</S.Ratio>
-          <S.Graph backgorundColor={color} width={teamData.ratio} />
-        </>
-      ) : (
-        <>
-          {isEnd ? (
-            <S.TeamBoxEndContent>
-              <S.ClassBox>
-                <S.Department>{teamData.department}</S.Department>
-                <S.Class fontSize={13}>{teamData.class}</S.Class>
-              </S.ClassBox>
-              <S.Score>{teamData.score}</S.Score>
-            </S.TeamBoxEndContent>
-          ) : (
-            <S.OfflineClassBox>
-              <S.Department>{teamData.department}</S.Department>
-              <S.Class fontSize={25}>{teamData.class}</S.Class>
-            </S.OfflineClassBox>
-          )}
-        </>
-      )}
-    </S.TeamBox>
-  );
-};
+import { Team } from "../../common/Team";
 
 interface ScheduleProps {
   scheduleData: {
@@ -96,9 +43,7 @@ export const GameSchedule: React.FC<ScheduleProps> = ({ scheduleData }) => {
         <p>{gameType}</p>
       </S.ScheduleTop>
       {isEnd ? (
-        <div style={{ width: "100%", fontSize: 17, fontWeight: 900 }}>
-          경기가 종료되었습니다.
-        </div>
+        <S.StatusDescription>경기가 종료되었습니다.</S.StatusDescription>
       ) : null}
       <S.Content isEnd={scheduleData.isEnd}>
         <Team
