+++ src/components/matchList/gameSchedule/styled.ts
@@ -53,6 +53,12 @@ export const ScheduleTop = styled.div`
   font-weight: 900;
 `;
 
+export const StatusDescription = styled.div`
+  width: 100%;
+  font-size: 17px;
+  font-weight: 900;
+`;
+
 export const Content = styled.div<ContentProps>`
   width: 100%;
   display: flex;
@@ -61,64 +67,8 @@ export const Content = styled.div<ContentProps>`
   margin-top: 10px;
 `;
 
-export const TeamBox = styled.div<TeamBoxProps>`
-  width: 100%;
-  font-weight: bold;
-  display: flex;
-  flex-direction: column;
-  align-items: ${(props) => props.alignItems};
-  color: ${(props) => props.color};
-  gap: 5px;
-`;
-
-export const TeamBoxEndContent = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 5px;
-`;
-
-export const ClassBox = styled.div`
-  display: flex;
-`;
-
-export const OfflineClassBox = styled.div`
-  text-align: center;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-`;
-
-export const Department = styled.p`
-  font-size: 13px;
-`;
-
-export const Class = styled.p<ClassProps>`
-  font-size: ${(props) => props.fontSize + "px"};
-`;
-
 export const IconBox = styled.span`
   font-size: 25px;
   color: ${colors.uiBrightGray};
   font-weight: bold;
-`;
-
-export const Ratio = styled.p`
-  font-size: 40px;
-`;
-
-export const Graph = styled.div<GraphProps>`
-  width: ${(props) => props.width}%;
-  height: 14px;
-  border-radius: 50px;
-  background-color: ${(props) => props.backgorundColor};
-`;
-
-export const ScoreBox = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-export const Score = styled.p`
-  font-size: 45px;
 `;
